refactor: :truck: tag, projectの型定義をそれぞれのファイルに移動(#28)
