+++ backend/golang/app/view/response/project.go
@@ -1,0 +1,6 @@
+package response
+
+type ProjectResponse struct {
+	Id    string `json:"id"`
+	Title string `json:"title"`
+}
