+++ backend/golang/app/view/response/tag.go
@@ -1,0 +1,6 @@
+package response
+
+type TagResponse struct {
+	Id   string `json:"id"`
+	Name string `json:"name"`
+}
