+++ backend/golang/app/view/response/todo.go
@@ -2,16 +2,6 @@ package response
 
 import "time"
 
-type ProjectResponse struct {
-	Id    string `json:"id"`
-	Title string `json:"title"`
-}
-
-type TagResponse struct {
-	Id   string `json:"id"`
-	Name string `json:"name"`
-}
-
 type TodosResponse struct {
 	Id      string          `json:"id"`
 	Name    string          `json:"name"`
